chore: ensures artifacts work
